requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=22.0.0
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0
redis>=5.2.0
celery>=5.4.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
httpx>=0.28.0
google-cloud-aiplatform>=1.74.0
google-cloud-storage>=2.19.0
google-cloud-bigquery>=3.20.0
boto3>=1.34.0
pydicom>=3.0.1
Pillow>=11.1.0
pandas>=2.2.0
openpyxl>=3.1.5
slowapi>=0.1.9
aiokafka>=0.10.0
pypdf>=4.0.0